requirements.txt
ollama>=0.1.7
Pillow>=10.0.0